feat(stark-build): merge webpack configurations into a single new webpack.config.js file
ISSUES CLOSED: #2558

BREAKING CHANGE:
  Simplify and improve integration of stark-build inside angular projects. Need to adapt "angular.json" file.

  Before:
  ```txt
  {
    //...
    "architect": {
      "build": {
        "builder": "@angular-builders/custom-webpack:browser",
        "options": {
          "customWebpackConfig": {
            "path": "./node_modules/@nationalbankbelgium/stark-build/config/webpack-partial.dev.js",
            "mergeStrategies": {
              "modules.rules": "prepend",
              "plugins": "prepend",
              "devServer": "prepend",
              "replaceDuplicatePlugins": false
            }
          },
          // ...
        },
        "configurations": {
          // ...
          "production": {
            "customWebpackConfig": {
              "path": "./node_modules/@nationalbankbelgium/stark-build/config/webpack-partial.prod.js",
              "mergeStrategies": {
                "modules.rules": "prepend",
                "plugins": "prepend",
                "replaceDuplicatePlugins": false
              }
            },
            // ...
          }
        }
      }
    }
  }
  ```

  After:
  ```txt
  {
    //...
    "architect": {
      "build": {
        "builder": "@angular-builders/custom-webpack:browser",
        "options": {
          "customWebpackConfig": {
            "path": "./node_modules/@nationalbankbelgium/stark-build/config/webpack.config.js"
            // mergeStrategies property is removed
          },
          // ...
        },
        "configurations": {
          // ...
          "production": {
            // "customWebpackConfig" property is removed
            // ...
          }
        }
      }
    }
  }
  ```
